first-project/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
python-wordpress-xmlrpc==2.3
jinja2==3.1.3
Pillow==10.2.0
